# backend/docs/api/openapi.pretty.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Niyat Ride API",
        "description": "API Documentation for Niyat Ride",
        "contact": {
            "name": "Niyat Support",
            "email": "backendteam@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/dispatcher/rides/{rideId}": {
            "get": {
                "tags": [
                    "Dispatcher Ride Creation"
                ],
                "summary": "Get specific ride details",
                "operationId": "getRideById",
                "parameters": [
                    {
                        "name": "rideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherRideResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/rides": {
            "get": {
                "tags": [
                    "Ride Management"
                ],
                "summary": "Get all rides with advanced filtering",
                "operationId": "getAllRides",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "requestedAt"
                        }
                    },
                    {
                        "name": "sortDirection",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "REQUESTED",
                                "ACCEPTED",
                                "IN_PROGRESS",
                                "COMPLETED",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "driverId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "customerId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "minAmount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "maxAmount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/rides/{id}": {
            "get": {
                "tags": [
                    "Ride Management"
                ],
                "summary": "Get specific ride details",
                "operationId": "getRideById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RideResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/rides/requests": {
            "get": {
                "tags": [
                    "Ride Management"
                ],
                "summary": "Get all ride requests with filtering",
                "operationId": "getAllRideRequests",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "requestedAt"
                        }
                    },
                    {
                        "name": "sortDirection",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "REQUESTED",
                                "ACCEPTED",
                                "IN_PROGRESS",
                                "COMPLETED",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "driverId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "customerId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "minAmount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "maxAmount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LocationDTO": {
                "required": [
                    "latitude",
                    "longitude"
                ],
                "type": "object",
                "properties": {
                    "latitude": {
                        "maximum": 90.0,
                        "exclusiveMaximum": false,
                        "minimum": -90.0,
                        "exclusiveMinimum": false,
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "maximum": 180.0,
                        "exclusiveMaximum": false,
                        "minimum": -180.0,
                        "exclusiveMinimum": false,
                        "type": "number",
                        "format": "double"
                    },
                    "address": {
                        "maxLength": 255,
                        "minLength": 0,
                        "type": "string"
                    },
                    "landmark": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "notes": {
                        "maxLength": 500,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "DispatcherRideResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerPhoneNumber": {
                        "type": "string"
                    },
                    "driverId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "driverName": {
                        "type": "string"
                    },
                    "vehicleTypeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "vehicleTypeName": {
                        "type": "string"
                    },
                    "pickupLocation": {
                        "$ref": "#/components/schemas/LocationDTO"
                    },
                    "dropoffLocation": {
                        "$ref": "#/components/schemas/LocationDTO"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "REQUESTED",
                            "ACCEPTED",
                            "IN_PROGRESS",
                            "COMPLETED",
                            "CANCELLED"
                        ]
                    },
                    "estimatedCost": {
                        "type": "number"
                    },
                    "finalCost": {
                        "type": "number"
                    },
                    "distanceKm": {
                        "type": "number",
                        "format": "double"
                    },
                    "estimatedDurationMin": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "nearbyDrivers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NearbyDriverDTO"
                        }
                    },
                    "basePrice": {
                        "type": "number"
                    },
                    "distancePrice": {
                        "type": "number"
                    },
                    "priceBreakdown": {
                        "type": "string"
                    },
                    "requestedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "acceptedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cancelledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cancellationReason": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "dispatcherId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "dispatcherName": {
                        "type": "string"
                    }
                }
            },
            "NearbyDriverDTO": {
                "type": "object",
                "properties": {
                    "driverId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "distanceKm": {
                        "type": "number",
                        "format": "double"
                    },
                    "estimatedArrivalMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "vehicle": {
                        "$ref": "#/components/schemas/VehicleInfoDTO"
                    },
                    "status": {
                        "type": "string"
                    },
                    "currentLatitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "currentLongitude": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
            "VehicleInfoDTO": {
                "type": "object",
                "properties": {
                    "vehicleType": {
                        "type": "string"
                    },
                    "vehicleModel": {
                        "type": "string"
                    },
                    "plateNumber": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "color": {
                        "type": "string"
                    }
                }
            },
...
            "RideResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "passengerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "passengerName": {
                        "type": "string"
                    },
                    "driverId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "driverName": {
                        "type": "string"
                    },
                    "vehicleTypeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "vehicleTypeName": {
                        "type": "string"
                    },
                    "pickupLatitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "pickupLongitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "pickupAddress": {
                        "type": "string"
                    },
                    "dropoffLatitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "dropoffLongitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "dropoffAddress": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "REQUESTED",
                            "ACCEPTED",
                            "IN_PROGRESS",
                            "COMPLETED",
                            "CANCELLED"
                        ]
                    },
                    "estimatedCost": {
                        "type": "number"
                    },
                    "finalCost": {
                        "type": "number"
                    },
                    "distanceKm": {
                        "type": "number",
                        "format": "double"
                    },
                    "estimatedDurationMin": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "requestedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "acceptedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cancelledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cancellationReason": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}